+++ clang/test/Misc/warning-wall.c
@@ -101,6 +101,6 @@ CHECK-NEXT:  -Wswitch
 CHECK-NEXT:  -Wswitch-bool
 CHECK-NEXT:  -Wmisleading-indentation
 CHECK-NEXT:  -Wpacked-non-pod
-
+CHECK-NEXT:  -Wvla-extension
 
 CHECK-NOT:-W